Cleanup folder structure and separate pages from components
