+++ src/components/Inputs/Checkbox.vue
