+++ src/components/SidebarPlugin/SideBar.vue
