+++ src/components/SidebarPlugin/SidebarLink.vue
