+++ src/components/SidebarPlugin/index.js
