+++ src/globalComponents.js
@@ -1,5 +1,5 @@
-import fgInput from './components/UIComponents/Inputs/formGroupInput.vue'
-import DropDown from './components/UIComponents/Dropdown.vue'
+import fgInput from './components/Inputs/formGroupInput.vue'
+import DropDown from './components/Dropdown.vue'
 
 /**
  * You can register global components here and use them as a plugin in your main Vue instance
