+++ src/layout/MobileMenu.vue
@@ -35,7 +35,7 @@
   </ul>
 </template>
 <script>
-  import DropDown from 'src/components/UIComponents/Dropdown.vue'
+  import DropDown from 'src/components/Dropdown.vue'
 
   export default {
     components: {
