+++ src/light-bootstrap-main.js
@@ -1,12 +1,12 @@
 import VTooltip from 'v-tooltip'
 // Notifications plugin
-import Notifications from 'src/components/UIComponents/NotificationPlugin'
+import Notifications from 'src/components/NotificationPlugin'
 // A plugin file where you could register global components used across the app
 import GlobalComponents from './globalComponents'
 // A plugin file where you could register global directives
 import GlobalDirectives from './globalDirectives'
 // Sidebar on the right. Used as a local plugin in DashboardLayout.vue
-import SideBar from './components/UIComponents/SidebarPlugin'
+import SideBar from './components/SidebarPlugin'
 
 // asset imports
 import 'bootstrap/dist/css/bootstrap.css'
