+++ src/pages/Notifications.vue
@@ -102,7 +102,7 @@
   </div>
 </template>
 <script>
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
+  import Card from 'src/components/Cards/Card.vue'
 
   export default {
     components: {
