+++ src/pages/Overview.vue
@@ -170,11 +170,11 @@
   </div>
 </template>
 <script>
-  import ChartCard from 'src/components/UIComponents/Cards/ChartCard.vue'
-  import StatsCard from 'src/components/UIComponents/Cards/StatsCard.vue'
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
-  import LTable from 'src/components/UIComponents/Table.vue'
-  import Checkbox from 'src/components/UIComponents/Inputs/Checkbox.vue'
+  import ChartCard from 'src/components/Cards/ChartCard.vue'
+  import StatsCard from 'src/components/Cards/StatsCard.vue'
+  import Card from 'src/components/Cards/Card.vue'
+  import LTable from 'src/components/Table.vue'
+  import Checkbox from 'src/components/Inputs/Checkbox.vue'
 
   export default {
     components: {
