+++ src/pages/TableList.vue
@@ -38,8 +38,8 @@
   </div>
 </template>
 <script>
-  import LTable from 'src/components/UIComponents/Table.vue'
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
+  import LTable from 'src/components/Table.vue'
+  import Card from 'src/components/Cards/Card.vue'
   const tableColumns = ['Id', 'Name', 'Salary', 'Country', 'City']
   const tableData = [{
     id: 1,
