+++ src/pages/Typography.vue
@@ -84,7 +84,7 @@
   </div>
 </template>
 <script>
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
+  import Card from 'src/components/Cards/Card.vue'
 
   export default {
     components: {
