+++ src/pages/UserProfile/EditProfileForm.vue
@@ -99,7 +99,7 @@
   </card>
 </template>
 <script>
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
+  import Card from 'src/components/Cards/Card.vue'
 
   export default {
     components: {
