+++ src/pages/UserProfile/UserCard.vue
@@ -22,7 +22,7 @@
   </card>
 </template>
 <script>
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
+  import Card from 'src/components/Cards/Card.vue'
   export default {
     components: {
       Card
